# server/internal/controller/transport/http/openapi/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Project Template API",
		"version": "0.1.0",
		"description": "Contract for the Project Template modular monolith."
	},
...
	"paths": {
...
		"/auth/csrf": {
			"get": {
				"operationId": "getCSRFToken",
				"summary": "Get CSRF token",
				"parameters": [],
				"responses": {
					"200": {
						"description": "The request has succeeded.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CSRFResponse"
								}
							}
						}
					},
					"401": {
						"description": "Access is unauthorized.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				},
				"tags": ["Auth"],
				"security": [
					{
						"SessionCookieAuth": []
					}
				]
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "loginUser",
				"summary": "Log in user",
				"parameters": [],
				"responses": {
					"200": {
						"description": "The request has succeeded.",
						"headers": {
							"Set-Cookie": {
								"required": true,
								"schema": {
									"type": "string"
								}
							}
						},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponse"
								}
							}
						}
					},
					"400": {
						"description": "The server could not understand the request due to invalid syntax.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"401": {
						"description": "Access is unauthorized.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"422": {
						"description": "Client error",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				},
				"tags": ["Auth"],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginRequest"
							}
						}
					}
				}
			}
		},
		"/auth/logout": {
			"post": {
				"operationId": "logoutUser",
				"summary": "Log out user",
				"parameters": [
					{
						"$ref": "#/components/parameters/CSRFHeader"
					}
				],
				"responses": {
					"204": {
						"description": "There is no content to send for this request, but the headers may be useful. ",
						"headers": {
							"Set-Cookie": {
								"required": true,
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"401": {
						"description": "Access is unauthorized.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"403": {
						"description": "Access is forbidden.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				},
				"tags": ["Auth"],
				"security": [
					{
						"SessionCookieAuth": []
					}
				]
			}
		},
		"/auth/me": {
			"get": {
				"operationId": "getCurrentUser",
				"summary": "Get current user",
				"parameters": [],
				"responses": {
					"200": {
						"description": "The request has succeeded.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponse"
								}
							}
						}
					},
					"401": {
						"description": "Access is unauthorized.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				},
				"tags": ["Auth"],
				"security": [
					{
						"SessionCookieAuth": []
					}
				]
			}
		},
		"/auth/register": {
			"post": {
				"operationId": "registerUser",
				"summary": "Register user",
				"parameters": [],
				"responses": {
					"201": {
						"description": "The request has succeeded and a new resource has been created as a result.",
						"headers": {
							"Set-Cookie": {
								"required": true,
								"schema": {
									"type": "string"
								}
							}
						},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponse"
								}
							}
						}
					},
					"400": {
						"description": "The server could not understand the request due to invalid syntax.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"409": {
						"description": "The request conflicts with the current state of the server.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"422": {
						"description": "Client error",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				},
				"tags": ["Auth"],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterRequest"
							}
						}
					}
				}
			}
		},
...
		"/workspaces": {
			"get": {
				"operationId": "listWorkspaces",
				"summary": "List workspaces",
				"parameters": [],
				"responses": {
					"200": {
						"description": "The request has succeeded.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WorkspaceListResponse"
								}
							}
						}
					},
					"401": {
						"description": "Access is unauthorized.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				},
				"tags": ["Workspaces"],
				"security": [
					{
						"SessionCookieAuth": []
					}
				]
			},
			"post": {
				"operationId": "createWorkspace",
				"summary": "Create workspace",
				"parameters": [
					{
						"$ref": "#/components/parameters/CSRFHeader"
					}
				],
				"responses": {
					"201": {
						"description": "The request has succeeded and a new resource has been created as a result.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WorkspaceResponse"
								}
							}
						}
					},
					"400": {
						"description": "The server could not understand the request due to invalid syntax.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"401": {
						"description": "Access is unauthorized.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"403": {
						"description": "Access is forbidden.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"422": {
						"description": "Client error",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				},
				"tags": ["Workspaces"],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateWorkspaceRequest"
							}
						}
					}
				},
				"security": [
					{
						"SessionCookieAuth": []
					}
				]
			}
		},
		"/workspaces/{workspaceId}": {
			"get": {
				"operationId": "getWorkspace",
				"summary": "Get workspace",
				"parameters": [
					{
						"$ref": "#/components/parameters/WorkspaceIdPath"
					}
				],
				"responses": {
					"200": {
						"description": "The request has succeeded.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WorkspaceResponse"
								}
							}
						}
					},
					"401": {
						"description": "Access is unauthorized.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"403": {
						"description": "Access is forbidden.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"404": {
						"description": "The server cannot find the requested resource.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"422": {
						"description": "Client error",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				},
				"tags": ["Workspaces"],
				"security": [
					{
						"SessionCookieAuth": []
					}
				]
			},
			"patch": {
				"operationId": "updateWorkspace",
				"summary": "Update workspace",
				"parameters": [
					{
						"$ref": "#/components/parameters/WorkspaceIdPath"
					},
					{
						"$ref": "#/components/parameters/CSRFHeader"
					}
				],
				"responses": {
					"200": {
						"description": "The request has succeeded.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WorkspaceResponse"
								}
							}
						}
					},
					"400": {
						"description": "The server could not understand the request due to invalid syntax.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"401": {
						"description": "Access is unauthorized.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"403": {
						"description": "Access is forbidden.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"404": {
						"description": "The server cannot find the requested resource.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"422": {
						"description": "Client error",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				},
				"tags": ["Workspaces"],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateWorkspaceRequest"
							}
						}
					}
				},
				"security": [
					{
						"SessionCookieAuth": []
					}
				]
			},
			"delete": {
				"operationId": "deleteWorkspace",
				"summary": "Delete workspace",
				"parameters": [
					{
						"$ref": "#/components/parameters/WorkspaceIdPath"
					},
					{
						"$ref": "#/components/parameters/CSRFHeader"
					}
				],
				"responses": {
					"204": {
						"description": "There is no content to send for this request, but the headers may be useful. "
					},
					"401": {
						"description": "Access is unauthorized.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"403": {
						"description": "Access is forbidden.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"404": {
						"description": "The server cannot find the requested resource.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"422": {
						"description": "Client error",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				},
				"tags": ["Workspaces"],
				"security": [
					{
						"SessionCookieAuth": []
					}
				]
			}
		},
		"/workspaces/{workspaceId}/members": {
			"get": {
				"operationId": "listWorkspaceMembers",
				"summary": "List workspace members",
				"parameters": [
					{
						"$ref": "#/components/parameters/WorkspaceIdPath"
					}
				],
				"responses": {
					"200": {
						"description": "The request has succeeded.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WorkspaceMemberListResponse"
								}
							}
						}
					},
					"401": {
						"description": "Access is unauthorized.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"403": {
						"description": "Access is forbidden.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"404": {
						"description": "The server cannot find the requested resource.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"422": {
						"description": "Client error",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				},
				"tags": ["Workspace Members"],
				"security": [
					{
						"SessionCookieAuth": []
					}
				]
			},
			"post": {
				"operationId": "addWorkspaceMember",
				"summary": "Add workspace member",
				"parameters": [
					{
						"$ref": "#/components/parameters/WorkspaceIdPath"
					},
					{
						"$ref": "#/components/parameters/CSRFHeader"
					}
				],
				"responses": {
					"201": {
						"description": "The request has succeeded and a new resource has been created as a result.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WorkspaceMemberResponse"
								}
							}
						}
					},
					"400": {
						"description": "The server could not understand the request due to invalid syntax.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"401": {
						"description": "Access is unauthorized.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"403": {
						"description": "Access is forbidden.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"404": {
						"description": "The server cannot find the requested resource.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"409": {
						"description": "The request conflicts with the current state of the server.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"422": {
						"description": "Client error",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				},
				"tags": ["Workspace Members"],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddWorkspaceMemberRequest"
							}
						}
					}
				},
				"security": [
					{
						"SessionCookieAuth": []
					}
				]
			}
		},
		"/workspaces/{workspaceId}/members/{userId}": {
			"patch": {
				"operationId": "updateWorkspaceMember",
				"summary": "Update workspace member",
				"parameters": [
					{
						"$ref": "#/components/parameters/WorkspaceIdPath"
					},
					{
						"$ref": "#/components/parameters/UserIdPath"
					},
					{
						"$ref": "#/components/parameters/CSRFHeader"
					}
				],
				"responses": {
					"200": {
						"description": "The request has succeeded.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WorkspaceMemberResponse"
								}
							}
						}
					},
					"400": {
						"description": "The server could not understand the request due to invalid syntax.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"401": {
						"description": "Access is unauthorized.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"403": {
						"description": "Access is forbidden.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"404": {
						"description": "The server cannot find the requested resource.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"409": {
						"description": "The request conflicts with the current state of the server.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"422": {
						"description": "Client error",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				},
				"tags": ["Workspace Members"],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateWorkspaceMemberRequest"
							}
						}
					}
				},
				"security": [
					{
						"SessionCookieAuth": []
					}
				]
			},
			"delete": {
				"operationId": "removeWorkspaceMember",
				"summary": "Remove workspace member",
				"parameters": [
					{
						"$ref": "#/components/parameters/WorkspaceIdPath"
					},
					{
						"$ref": "#/components/parameters/UserIdPath"
					},
					{
						"$ref": "#/components/parameters/CSRFHeader"
					}
				],
				"responses": {
					"204": {
						"description": "There is no content to send for this request, but the headers may be useful. "
					},
					"401": {
						"description": "Access is unauthorized.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"403": {
						"description": "Access is forbidden.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"404": {
						"description": "The server cannot find the requested resource.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"409": {
						"description": "The request conflicts with the current state of the server.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"422": {
						"description": "Client error",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				},
				"tags": ["Workspace Members"],
				"security": [
					{
						"SessionCookieAuth": []
					}
				]
			}
		},
		"/workspaces/{workspaceId}/tasks": {
			"get": {
				"operationId": "listTasks",
				"summary": "List tasks",
				"parameters": [
					{
						"$ref": "#/components/parameters/WorkspaceIdPath"
					},
					{
						"$ref": "#/components/parameters/TaskListQuery"
					}
				],
				"responses": {
					"200": {
						"description": "The request has succeeded.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TaskListResponse"
								}
							}
						}
					},
					"401": {
						"description": "Access is unauthorized.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"403": {
						"description": "Access is forbidden.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"404": {
						"description": "The server cannot find the requested resource.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"422": {
						"description": "Client error",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				},
				"tags": ["Tasks"],
				"security": [
					{
						"SessionCookieAuth": []
					}
				]
			},
			"post": {
				"operationId": "createTask",
				"summary": "Create task",
				"parameters": [
					{
						"$ref": "#/components/parameters/WorkspaceIdPath"
					},
					{
						"$ref": "#/components/parameters/CSRFHeader"
					}
				],
				"responses": {
					"201": {
						"description": "The request has succeeded and a new resource has been created as a result.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TaskResponse"
								}
							}
						}
					},
					"400": {
						"description": "The server could not understand the request due to invalid syntax.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"401": {
						"description": "Access is unauthorized.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"403": {
						"description": "Access is forbidden.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"404": {
						"description": "The server cannot find the requested resource.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"422": {
						"description": "Client error",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				},
				"tags": ["Tasks"],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTaskRequest"
							}
						}
					}
				},
				"security": [
					{
						"SessionCookieAuth": []
					}
				]
			}
		},
...
	},
	"components": {
		"parameters": {
			"CSRFHeader": {
				"name": "X-CSRF-Token",
				"in": "header",
				"required": true,
				"schema": {
					"type": "string",
					"minLength": 1
				}
			},
...
		},
		"schemas": {
...
			"AuthResponse": {
				"type": "object",
				"required": ["user"],
				"properties": {
					"user": {
						"$ref": "#/components/schemas/User"
					}
				},
				"examples": [
					{
						"user": {
							"id": "11111111-1111-1111-1111-111111111111",
							"email": "owner@example.com",
							"displayName": "Alex Morgan",
							"createdAt": "2026-07-10T08:00:00Z",
							"updatedAt": "2026-07-10T08:00:00Z"
						}
					}
				]
			},
			"CSRFResponse": {
				"type": "object",
				"required": ["token"],
				"properties": {
					"token": {
						"type": "string",
						"format": "password"
					}
				},
				"examples": [
					{
						"token": "a-one-time-readable-csrf-token"
					}
				]
			},
...
			"LoginRequest": {
				"type": "object",
				"required": ["email", "password"],
				"properties": {
					"email": {
						"$ref": "#/components/schemas/email"
					},
					"password": {
						"type": "string",
						"minLength": 12,
						"maxLength": 128,
						"format": "password"
					}
				},
				"examples": [
					{
						"email": "owner@example.com",
						"password": "correct-horse-battery-staple"
					}
				]
			},
			"ProblemDetails": {
				"type": "object",
				"required": ["type", "title", "status", "code"],
				"properties": {
					"type": {
						"type": "string",
						"format": "uri"
					},
					"title": {
						"type": "string"
					},
					"status": {
						"type": "integer",
						"format": "int32"
					},
					"code": {
						"type": "string",
						"enum": [
							"BAD_REQUEST",
							"UNAUTHORIZED",
							"FORBIDDEN",
							"NOT_FOUND",
							"CONFLICT",
							"VALIDATION_FAILED",
							"INTERNAL_ERROR",
							"SERVICE_UNAVAILABLE",
							"METHOD_NOT_ALLOWED",
							"AUTH_MISSING_SESSION",
							"AUTH_INVALID_SESSION",
							"AUTH_INVALID_CREDENTIALS",
							"AUTH_INVALID_CSRF",
							"INSUFFICIENT_ROLE",
							"EMAIL_ALREADY_EXISTS",
							"WORKSPACE_NOT_FOUND",
							"WORKSPACE_MEMBER_NOT_FOUND",
							"WORKSPACE_MEMBER_ALREADY_EXISTS",
							"WORKSPACE_LAST_OWNER",
							"TASK_ASSIGNEE_NOT_MEMBER",
							"TASK_NOT_FOUND"
						]
					},
					"detail": {
						"type": "string"
					},
					"requestId": {
						"type": "string"
					},
					"errors": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/ProblemError"
						}
					}
				},
				"examples": [
					{
						"type": "https://project-template.example/problems/validation-failed",
						"title": "Validation failed",
						"status": 422,
						"code": "VALIDATION_FAILED",
						"detail": "One or more request fields are invalid.",
						"requestId": "req_01J00000000000000000000000",
						"errors": [
							{
								"code": "REQUIRED",
								"message": "title is required",
								"location": "body.title"
							}
						]
					}
				]
			},
...
		},
		"securitySchemes": {
			"SessionCookieAuth": {
				"type": "apiKey",
				"in": "cookie",
				"name": "__Host-project_template_session"
			}
		}
	},
	"servers": [
		{
			"url": "/api/v1",
			"description": "Version 1 API",
			"variables": {}
		}
	]
}